Move utils to compilation_pipeline.py
